Plume-WSS-main/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
bidict==0.22.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.6
clickclick==20.10.2
connexion==2.14.2
dataclasses-json==0.5.13
distro==1.9.0
ecdsa==0.18.0
faiss-cpu==1.7.4
Flask==2.2.5
Flask-Cors==4.0.0
flask-marshmallow==0.15.0
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.0.5
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==6.8.0
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
langchain==0.0.334
langsmith==0.0.62
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.1.0
numexpr==2.8.4
numpy==1.25.1
openai==1.10.0
openapi-schema-pydantic==1.2.4
packaging==23.1
pluggy==1.3.0
priomptipy==0.16
protobuf==4.21.12
pyasn1==0.5.0
pydantic==1.10.11
pytest==7.4.3
pytest-subtests==0.11.0
python-dotenv==1.0.0
python-engineio==4.5.1
python-jose==3.3.0
python-socketio==5.8.0
PyYAML==6.0.1
referencing==0.30.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.9.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
tenacity==8.2.2
tiktoken==0.5.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.1.0
websockets==11.0.3
Werkzeug==2.2.3
yarl==1.9.2
zipp==3.16.2
